requirements.txt
# Flask Web Framework
flask==3.0.0
flask-cors==4.0.0
werkzeug==3.0.1

# CLIP Model
transformers>=4.36.2
torch>=2.2.0
torchvision>=0.17.0

# Image Processing
Pillow>=10.1.0
numpy<2  # Required for PyTorch compatibility
opencv-python>=4.8.0  # HDR merging, image alignment
rawpy>=0.18.0  # RAW file support (CR2, NEF, ARW)

# SDXL + ControlNet
diffusers>=0.25.0
accelerate>=0.25.0
safetensors>=0.4.0

# Utilities
python-dotenv==1.0.0

